+++ src/invoices/InvoicesList/hooks/useBuildQuery/useBuildQuery.js
@@ -44,6 +44,7 @@ export const useBuildQuery = () => {
       [FILTERS.FUND_CODE]: buildArrayFieldQuery.bind(null, ['invoiceLines.fundDistributions']),
       [FILTERS.EXPENSE_CLASS]: buildArrayFieldQuery.bind(null, ['invoiceLines.fundDistributions']),
       [FILTERS.LOCK_TOTAL]: buildNumberRangeQuery.bind(null, [FILTERS.LOCK_TOTAL]),
+      [FILTERS.TOTAL_AMOUNT]: buildNumberRangeQuery.bind(null, [FILTERS.TOTAL_AMOUNT]),
     },
   ), []);
 };